refactoring SV to make room for onnx impl
